Доработка защиты от дурака

+++ life_designer/ViewModel/Del_categoryViewModel.cs
@@ -28,23 +28,41 @@ namespace life_designer.ViewModel
             }
         }
 
+        private string errText;
+        public string ErrText
+        {
+            get { return errText; }
+            set
+            {
+                errText = value;
+                OnPropertyChanged("ErrText");
+            }
+        }
+
         public ICommand RemoveCategoryCommand { get; private set; }
 
  
         private void RemoveCategory(object parameter)
         {
-            using (var context = new DataBaseContext())
+            if (Text == null || Text == "")
             {
-                var category = context.Categorys.Where(c => c.Name == Text).ExecuteDelete();
-                foreach (var coll in ItemsCollection.Items)
+                ErrText = "Обязательно для заполнения";
+            }
+            else
+            {
+                using (var context = new DataBaseContext())
                 {
-                    if (coll.Header == Text)
+                    var category = context.Categorys.Where(c => c.Name == Text).ExecuteDelete();
+                    foreach (var coll in ItemsCollection.Items)
                     {
-                        ItemsCollection.Items.Remove(coll);
-                        break;
+                        if (coll.Header == Text)
+                        {
+                            ItemsCollection.Items.Remove(coll);
+                            break;
+                        }
                     }
+                    CloseWindowCommand.Execute(null);
                 }
-                CloseWindowCommand.Execute(null);
             }
         }
